+++ libc/sysv/consts/lock.h
@@ -4,11 +4,11 @@
 #if !(__ASSEMBLER__ + __LINKER__ + 0)
 COSMOPOLITAN_C_START_
 
-extern const long LOCK_EX;
-extern const long LOCK_NB;
-extern const long LOCK_SH;
-extern const long LOCK_UN;
-extern const long LOCK_UNLOCK_CACHE;
+extern const int LOCK_EX;
+extern const int LOCK_NB;
+extern const int LOCK_SH;
+extern const int LOCK_UN;
+extern const int LOCK_UNLOCK_CACHE;
 
 COSMOPOLITAN_C_END_
 #endif /* !(__ASSEMBLER__ + __LINKER__ + 0) */
